Render extension detail as separate page
