+++ app/src/modules/settings/index.ts
@@ -13,8 +13,8 @@ import Extensions from './routes/extensions/extensions.vue';
 import FlowOperationDetail from './routes/flows/components/operation-detail.vue';
 import FlowsDetail from './routes/flows/flow.vue';
 import FlowsOverview from './routes/flows/overview.vue';
+import MarketplaceExtension from './routes/marketplace/extension.vue';
 import MarketplaceRegistry from './routes/marketplace/registry.vue';
-import MarketplaceExtension from './routes/marketplace/routes/extension.vue';
 import NotFound from './routes/not-found.vue';
 import PresetsCollection from './routes/presets/collection/collection.vue';
 import PresetsItem from './routes/presets/item.vue';
@@ -232,10 +232,15 @@ export default defineModule({
 			component: Extensions,
 		},
 		{
-			name: 'marketplace-registry',
+			name: 'marketplace',
 			path: 'marketplace',
-			component: MarketplaceRegistry,
+			component: RouterPass,
 			children: [
+				{
+					name: 'marketplace-registry',
+					path: '',
+					component: MarketplaceRegistry,
+				},
 				{
 					name: 'marketplace-extension',
 					path: ':extensionId',
